+++ oomact/include/aslam/calibration/calibrator/SetActivator.h
@@ -4,7 +4,7 @@
 #include <iosfwd>
 #include <unordered_set>
 
-#include <aslam/calibration/CalibrationConfI.h>
+#include <aslam/calibration/calibrator/CalibrationConfI.h>
 
 namespace aslam {
 namespace calibration {
