+++ oomact/include/aslam/calibration/model/sensors/Imu.h
@@ -4,7 +4,7 @@
 #include <aslam/calibration/data/MeasurementsContainer.h>
 #include <aslam/calibration/input/InputReceiverI.h>
 #include <aslam/calibration/model/Module.h>
-#include <aslam/calibration/model/StateCarrier.h>
+#include <aslam/calibration/calibrator/StateCarrier.h>
 #include <aslam/calibration/model/Sensor.hpp>
 
 namespace aslam {
