+++ oomact/src/calibrator/BatchCalibrator.cpp
@@ -6,9 +6,9 @@
 
 #include <aslam/calibration/calibrator/AbstractCalibrator.h>
 #include "aslam/calibration/calibrator/CalibratorI.hpp"
-#include <aslam/calibration/CalibrationProblem.hpp>
+#include <aslam/calibration/calibrator/CalibrationProblem.hpp>
 #include <aslam/calibration/data/MapStorage.h>
-#include <aslam/calibration/model/StateCarrier.h>
+#include <aslam/calibration/calibrator/StateCarrier.h>
 #include <aslam/calibration/model/Sensor.hpp>
 
 namespace aslam {
